Implemented new cookie form into relevant types
